03_litellm-agent/pyproject.toml
[project]
name = "litellm-gemini-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.72.0",
    "python-dotenv>=1.1.0",
]
